Practicamente terminados los menus. Comienzo con el debugging

+++ Cliente.h
@@ -40,7 +40,7 @@ private:
     int _idCliente;
     char _nombre[50];
     char _apellido[50];
-    int _cuil;
+    long long _cuil;
     int _tipoCliente;
     int _numeroTelefono;
     char _email[20];
